requirements.txt
# Web and data handling
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3

# Storage
minio==7.1.15
pymongo==4.4.1
python-dotenv==1.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.25.0

# Data processing
numpy==1.24.3 